Added Template Selection UI
